functional tests and int values refactor

+++ internal/models/driver_type.go
@@ -1,7 +1,7 @@
 package models
 
 type DriverType struct {
-	ID    int    `json:"id" bson:"id"`
+	ID    uint16 `json:"id" bson:"id"`
 	Name  string `json:"name" bson:"name"`
-	Value int    `json:"value" bson:"value"`
+	Value uint16 `json:"value" bson:"value"`
 }
